2022.05.04

+++ Clones/Wetube/wetube/src/controllers/videoController.js
@@ -19,32 +19,52 @@ export const watch = async (req, res) => {
 };
 
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
+
   const video = await Video.findById(id);
 
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
-  } else {
-    return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
   }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+
+  return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
-  const { id } = req.params;
-  const { title, description, hashtags } = req.body;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+    body: { title, description, hashtags },
+  } = req;
+
   const video = await Video.exists({ _id: id });
 
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
-  } else {
-    await Video.findByIdAndUpdate(id, {
-      title,
-      description,
-      hashtags: Video.formatHashtags(hashtags),
-    });
-
-    return res.redirect(`/videos/${id}`);
   }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+
+  await Video.findByIdAndUpdate(id, {
+    title,
+    description,
+    hashtags: Video.formatHashtags(hashtags),
+  });
+
+  return res.redirect(`/videos/${id}`);
 };
 
 export const getUpload = (req, res) => {
@@ -84,7 +104,22 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
+
+  const video = await Video.findById(id);
+
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
 
   await Video.findByIdAndDelete(id);
